add basic navigation tabs and adapt ParallaxScrollView and ThemedView to use default values instead of customizing everytime
