+++ mensasverige/app/(tabs)/(home)/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from 'expo-router';
+import React from 'react';
+
+export default function TabLayout() {
+
+  return (
+      <Stack>
+        <Stack.Screen name="index" options={{ headerShown: false }} />
+      </Stack>
+  );
+}
