+++ mensasverige/app/login.tsx
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+import ParallaxScrollView from '@/components/ParallaxScrollView';
+import { ThemedText } from '@/components/ThemedText';
+import { ThemedView } from '@/components/ThemedView';
+
+export default function LoginScreen() {
+  return (
+    <ParallaxScrollView>
+      <ThemedView >
+        <ThemedText type="title">Login!</ThemedText>
+      </ThemedView>
+    </ParallaxScrollView>
+  );
+}
+
+const styles = StyleSheet.create({
+
+});
